+++ tenantmanagercontrolplane/get_assigned_subscription_request_response.go
@@ -18,7 +18,7 @@ import (
 // Click https://docs.cloud.oracle.com/en-us/iaas/tools/go-sdk-examples/latest/tenantmanagercontrolplane/GetAssignedSubscription.go.html to see an example of how to use GetAssignedSubscriptionRequest.
 type GetAssignedSubscriptionRequest struct {
 
-	// OCID of the assigned subscription.
+	// OCID of the assigned Oracle Cloud Subscription.
 	AssignedSubscriptionId *string `mandatory:"true" contributesTo:"path" name:"assignedSubscriptionId"`
 
 	// The client request ID for tracing.