+++ tenantmanagercontrolplane/tenantmanagercontrolplane_domaingovernance_client.go
@@ -57,7 +57,7 @@ func newDomainGovernanceClientFromBaseClient(baseClient common.BaseClient, confi
 	common.ConfigCircuitBreakerFromGlobalVar(&baseClient)
 
 	client = DomainGovernanceClient{BaseClient: baseClient}
-	client.BasePath = "20200801"
+	client.BasePath = "20230401"
 	err = client.setConfigurationProvider(configProvider)
 	return
 }
@@ -141,7 +141,7 @@ func (client DomainGovernanceClient) createDomainGovernance(ctx context.Context,
 	defer common.CloseBodyIfValid(httpResponse)
 	response.RawResponse = httpResponse
 	if err != nil {
-		apiReferenceLink := "https://docs.oracle.com/iaas/api/#/en/organizations/20200801/DomainGovernance/CreateDomainGovernance"
+		apiReferenceLink := "https://docs.oracle.com/iaas/api/#/en/organizations/20230401/DomainGovernance/CreateDomainGovernance"
 		err = common.PostProcessServiceError(err, "DomainGovernance", "CreateDomainGovernance", apiReferenceLink)
 		return response, err
 	}
@@ -198,7 +198,7 @@ func (client DomainGovernanceClient) deleteDomainGovernance(ctx context.Context,
 	defer common.CloseBodyIfValid(httpResponse)
 	response.RawResponse = httpResponse
 	if err != nil {
-		apiReferenceLink := "https://docs.oracle.com/iaas/api/#/en/organizations/20200801/DomainGovernance/DeleteDomainGovernance"
+		apiReferenceLink := "https://docs.oracle.com/iaas/api/#/en/organizations/20230401/DomainGovernance/DeleteDomainGovernance"
 		err = common.PostProcessServiceError(err, "DomainGovernance", "DeleteDomainGovernance", apiReferenceLink)
 		return response, err
 	}
@@ -255,7 +255,7 @@ func (client DomainGovernanceClient) getDomainGovernance(ctx context.Context, re
 	defer common.CloseBodyIfValid(httpResponse)
 	response.RawResponse = httpResponse
 	if err != nil {
-		apiReferenceLink := "https://docs.oracle.com/iaas/api/#/en/organizations/20200801/DomainGovernance/GetDomainGovernance"
+		apiReferenceLink := "https://docs.oracle.com/iaas/api/#/en/organizations/20230401/DomainGovernance/GetDomainGovernance"
 		err = common.PostProcessServiceError(err, "DomainGovernance", "GetDomainGovernance", apiReferenceLink)
 		return response, err
 	}
@@ -312,7 +312,7 @@ func (client DomainGovernanceClient) listDomainGovernances(ctx context.Context, 
 	defer common.CloseBodyIfValid(httpResponse)
 	response.RawResponse = httpResponse
 	if err != nil {
-		apiReferenceLink := "https://docs.oracle.com/iaas/api/#/en/organizations/20200801/DomainGovernance/ListDomainGovernances"
+		apiReferenceLink := "https://docs.oracle.com/iaas/api/#/en/organizations/20230401/DomainGovernance/ListDomainGovernances"
 		err = common.PostProcessServiceError(err, "DomainGovernance", "ListDomainGovernances", apiReferenceLink)
 		return response, err
 	}
@@ -369,7 +369,7 @@ func (client DomainGovernanceClient) updateDomainGovernance(ctx context.Context,
 	defer common.CloseBodyIfValid(httpResponse)
 	response.RawResponse = httpResponse
 	if err != nil {
-		apiReferenceLink := "https://docs.oracle.com/iaas/api/#/en/organizations/20200801/DomainGovernance/UpdateDomainGovernance"
+		apiReferenceLink := "https://docs.oracle.com/iaas/api/#/en/organizations/20230401/DomainGovernance/UpdateDomainGovernance"
 		err = common.PostProcessServiceError(err, "DomainGovernance", "UpdateDomainGovernance", apiReferenceLink)
 		return response, err
 	}